+++ src/Datiss.Budget.Services/NHCoService.cs
@@ -210,13 +210,21 @@ namespace Datiss.Budget.Services
                                    .Where(_ => _.OrganizationId == organizationId)
                                    .Where(_ => _.ActivityType == activityType)
                                    .ToListAsync();
-            var childrens = await getChildren(organizationId, yearId, activityType);
+
+            IEnumerable<NHCo> childrens = new NHCo[] { };
+
+            if (organization.Type == OrganizationType.County || organization.Type == OrganizationType.Root)
+            {
+                childrens = await getChildren(organizationId, yearId, activityType);
+            }
+
 
             if (self.Count() == 0 && childrens.Count() == 0)
                 throw new DeleteNullRecordException();
             _dbSet.RemoveRange(self);
 
-            _dbSet.RemoveRange(childrens);
+            if (childrens.Any())
+                _dbSet.RemoveRange(childrens);
 
             var result = new OrganizationDeleteDataResult
             {
